[ADF-1761] removed old search component replaced with the new implementation
