--- ng2-components/ng2-alfresco-search/src/components/search-autocomplete.component.html
@@ -1,8 +1,0 @@
-<div class="adf-search-result" role="listbox" [id]="id" [ngClass]="_classList" #panel>
-    <ng-template
-        ngFor [ngForOf]="results"
-        [ngForTemplate]="template">
-    </ng-template>
-</div>
-
-
