--- ng2-components/ng2-alfresco-search/src/components/search-autocomplete.component.scss
@@ -1,86 +1,0 @@
-@mixin adf-search-autocomplete-theme($theme) {
-    $primary: map-get($theme, primary);
-    $accent: map-get($theme, accent);
-    $warn: map-get($theme, warn);
-    $foreground: map-get($theme, foreground);
-    $background: map-get($theme, background);
-    $mat-menu-border-radius: 2px !default;
-
-    .adf {
-
-        &-search-hide {
-            display: none;
-        }
-
-        &-search-show {
-            display: block;
-        }
-
-        &--search-result-container{
-            z-index: 5;
-        }
-
-        &-search-result {
-            @include mat-menu-base(2);
-            transform-origin: top left;
-            position: absolute;
-            z-index: 5;
-            color: mat-color($foreground, text);
-            background-color: mat-color($background, card);
-            margin: -20px 0 0 0;
-            border-radius: $mat-menu-border-radius;
-            border-collapse: collapse;
-            white-space: nowrap;
-            font-size: 14px;
-
-            a {
-                color: mat-color($foreground, text);
-                text-decoration: none;
-            }
-
-            table {
-                width: 300px;
-            }
-
-            tbody tr {
-                height: 32px;
-
-                &:hover {
-                    background-color: mat-color($background, hover);
-                }
-            }
-
-            td {
-                height: 32px;
-                padding: 8px;
-                text-align: left;
-                border-top: none;
-                border-bottom: none;
-            }
-
-            .highlight {
-                color: mat-color($primary, 900);
-            }
-
-            .img-td {
-                width: 30px;
-            }
-
-            .truncate {
-                width: 240px;
-                white-space: nowrap;
-                overflow: hidden;
-                text-overflow: ellipsis;
-            }
-
-            @media screen and (max-width: 400px) {
-                :host {
-                    right: 0;
-                }
-                .truncate {
-                    width: 200px;
-                }
-            }
-        }
-    }
-}
