+++ ng2-components/ng2-alfresco-search/src/components/search.component.scss
@@ -1,37 +1,48 @@
-:host .adf-data-table caption {
-    margin: 0 0 16px 0;
-    text-align: left;
-}
-:host .adf-data-table td {
-    white-space: nowrap;
-}
-:host .adf-data-table td.col-mimetype-icon {
-    width: 24px;
-}
-:host .col-display-name {
-    min-width: 250px;
-    overflow: hidden;
-    text-overflow: ellipsis;
-}
+@mixin adf-search-autocomplete-theme($theme) {
+    $primary: map-get($theme, primary);
+    $accent: map-get($theme, accent);
+    $warn: map-get($theme, warn);
+    $foreground: map-get($theme, foreground);
+    $background: map-get($theme, background);
+    $mat-menu-border-radius: 2px !default;
 
-.no-result-message {
-    height: 32px;
-    opacity: 0.26;
-    font-size: 24px;
-    line-height: 1.33;
-    letter-spacing: -1px;
-    color: #000000;
-}
+    .adf {
 
-.no-result__empty_doc_lib {
-    width: 565px;
-    height: 161px;
-    object-fit: contain;
-    margin-top: 17px;
-}
+        &-search-hide {
+            visibility: hidden;
+        }
 
-.empty_template {
-    text-align: center;
-    margin-top: 20px;
-    margin-bottom: 20px;
+        &-search-show {
+            visibility: visible;
+        }
+
+        &-search-result-autocomplete {
+            @include mat-menu-base(2);
+            transform-origin: top left;
+            position: absolute;
+            overflow: inherit;
+            max-width: 300px;
+            z-index: 5;
+            color: mat-color($foreground, text);
+            background-color: mat-color($background, card);
+            margin: -14px 0 0 14px;
+            border-radius: $mat-menu-border-radius;
+            border-collapse: collapse;
+            white-space: nowrap;
+            font-size: 14px;
+            a {
+              color: mat-color($foreground, text);
+              text-decoration: none;
+            }
+            .highlight {
+              color: mat-color($primary, 900);
+            }
+
+            @media screen and (max-width: 400px) {
+                width: 200px;
+                white-space: nowrap;
+                text-overflow: ellipsis;
+            }
+          }
+    }
 }
