+++ ng2-components/ng2-alfresco-search/src/components/search.old.component.scss
@@ -1,0 +1,37 @@
+:host .adf-data-table caption {
+    margin: 0 0 16px 0;
+    text-align: left;
+}
+:host .adf-data-table td {
+    white-space: nowrap;
+}
+:host .adf-data-table td.col-mimetype-icon {
+    width: 24px;
+}
+:host .col-display-name {
+    min-width: 250px;
+    overflow: hidden;
+    text-overflow: ellipsis;
+}
+
+.no-result-message {
+    height: 32px;
+    opacity: 0.26;
+    font-size: 24px;
+    line-height: 1.33;
+    letter-spacing: -1px;
+    color: #000000;
+}
+
+.no-result__empty_doc_lib {
+    width: 565px;
+    height: 161px;
+    object-fit: contain;
+    margin-top: 17px;
+}
+
+.empty_template {
+    text-align: center;
+    margin-top: 20px;
+    margin-bottom: 20px;
+}
