+++ ng2-components/ng2-alfresco-search/styles/index.scss
@@ -1,4 +1,4 @@
-@import '../src/components/search-autocomplete.component';
+@import '../src/components/search.component';
 @import '../src/components/search-control.component';
 
 @mixin alfresco-search-theme($theme) {
